+++ src/ui/repl.js
@@ -2,6 +2,7 @@ import blessed from '@hp4k1h5/blessed'
 
 import { evaluate } from './evaluate.js'
 import { intro } from './help.js'
+import { handleErr } from '../util/error.js'
 import { exit } from './evaluate.js'
 
 export function buildRepl(ws, options) {
@@ -70,7 +71,11 @@ export function buildRepl(ws, options) {
     ws.input.clearValue()
 
     // parse and handle input
-    await evaluate(ws, data)
+    try {
+      await evaluate(ws, data)
+    } catch (e) {
+      handleErr(ws, e)
+    }
 
     // wait to focus until evaluation completes
     ws.input.focus()
@@ -78,5 +83,7 @@ export function buildRepl(ws, options) {
 
   ws.input.on('focus', function () {
     ws.prevFocus.box.style.border = { fg: '#fc5' }
+    output.setFront()
+    ws.input.setFront()
   })
 }