+++ src/util/defaults.js
@@ -42,9 +42,19 @@ const list = {
   listTypes: ['mostactive', 'gainers', 'losers', 'iexvolume', 'iexpercent'],
 }
 
+const sectors = {
+  type: 'sectors',
+  yxhw: [0, 9, 4, 3],
+}
+
 const account = {
   type: 'account',
   yxhw: [0, 0, 12, 9],
+}
+
+const help = {
+  type: 'help',
+  yxhw: [1, 1, 9, 9],
 }
 
 export const defaults = {
@@ -55,5 +65,7 @@ export const defaults = {
   news,
   profile,
   list,
+  sectors,
   account,
+  help,
 }